Update and rename FAQs.html to Manual.html
Hi Robin & Nikolai - I've made a start, renamed it the user manual and changed the intro text. Could you update?
It would also be really helpful to have the separate FAQs (which will be changed to chapters) separately editable as the amount of html on one page is making my head hurt. Is it possible to create separate sections for the different chapters and have this page call them? Thanks and let me know if I'm doing this wrong :)