Remove global variables on repopulate_region
Repopulate region should happen when we load a new
regions data.  Setting it as a global variable was
causing it to work but a more measured approach is to
set it to false when we move region and set it to true
when we have loaded the data for the new region.